+++ src/app/core/+store/app.state.ts
@@ -1,0 +1,7 @@
+import { ProductsState } from './products';
+import { CartState } from './cart';
+
+export interface AppState {
+  products: ProductsState;
+  cart: CartState;
+}