+++ src/app/core/+store/cart/cart.state.ts
@@ -1,0 +1,9 @@
+import { Cart } from './../../../cart/models/cart.model';
+
+export interface CartState {
+  data: ReadonlyArray<Cart>;
+}
+
+export const initialCartState: CartState = {
+  data: []
+};