+++ src/app/core/+store/products/products.state.ts
@@ -1,0 +1,11 @@
+import { Product } from './../../../products/models/product.model';
+
+export interface ProductsState {
+  data: ReadonlyArray<Product>;
+  readonly error: Error | string;
+}
+
+export const initialProductsState: ProductsState = {
+  data: [],
+  error: null
+};