+++ src/app/products/components/product-list/product-list.component.ts
@@ -1,10 +1,13 @@
 import { Component, OnInit, Input } from '@angular/core';
 import { Router } from '@angular/router';
 
-import { Subscription} from 'rxjs';
+import { Subscription, Observable } from 'rxjs';
+
+import { Store, select } from '@ngrx/store';
+import { AppState, ProductsState } from './../../../core/+store';
+import * as ProductsActions from './../../../core/+store/products/products.actions';
 
 import { Product } from '../../models';
-import { ProductsPromiseService } from '../../services';
 import { CartObservableService } from '../../../cart';
 
 import { AutoUnsubscribe } from '../../../core';
@@ -19,19 +22,24 @@ import { AutoUnsubscribe } from '../../../core';
 export class ProductListComponent implements OnInit {
   @Input() mode?:string;
 
+  productsState$: Observable<ProductsState>;
+
   private sub: Subscription;
   private sub2: Subscription;
   private sub3: Subscription;
 
-  productList = this.productsPromiseService.getProducts();
-
   constructor(
-    private productsPromiseService: ProductsPromiseService,
     private cartObservableService: CartObservableService,
     private router: Router,
+    private store: Store<AppState>
   ) { }
 
   ngOnInit() {
+    console.log('Store works', this.store);
+
+    this.productsState$ = this.store.pipe(select('products'));
+    this.store.dispatch(new ProductsActions.GetProducts());
+
     this.mode = this.mode == 'admin' ? 'admin' : 'guest';
   }
 
@@ -61,17 +69,12 @@ export class ProductListComponent implements OnInit {
   }
 
   onEditProduct(product) {
-    //create absolute path /edit/productID
     const link = ['/edit', product.id];
     this.router.navigate(link);
   }
 
   onDeleteProduct(product) {
-    this.productsPromiseService.deleteProduct(product)
-      .then(() => {
-        this.productList = this.productsPromiseService.getProducts()
-      })
-      .catch(err => console.log(err));
+    this.store.dispatch(new ProductsActions.DeleteProduct(product));
   }
 
   onCreateProduct(product) {